feat(preset): cleanup require()'s and consolidate files
per feedback

BREAKING CHANGE: Removed unused Preset registry functions from LayerPresetManager. Please use
LayerPresetManager.registerService() instead.
