--- src/os/layer/preset/settingkey.js
@@ -1,14 +1,0 @@
-goog.module('os.layer.preset.SettingKey');
-goog.module.declareLegacyNamespace();
-
-const {BASE_KEY} = goog.require('os.layer.preset');
-
-
-/**
- * Settings keys for layer presets.
- * @enum {string}
- */
-exports = {
-  APPLIED_DEFAULTS: BASE_KEY + 'appliedDefaults',
-  PRESETS: BASE_KEY + 'presets'
-};
