+++ views/layer/vector.html
@@ -12,7 +12,7 @@
         </div>
         <div class="d-flex col">
           <select class="custom-select" ng-model="vector.preset"
-              ng-options="(preset.label + (preset.default === true ? ' (default)' : '') + (preset.published === true ? '' : ' (unpublished)')) for preset in vector.presets track by preset.id">
+              ng-options="((preset.published === true ? '' : '*') + preset.label + (preset.default === true ? ' (default)' : '')) for preset in vector.presets track by preset.id">
           </select>
           <button type="button" class="btn btn-primary col-auto ml-1" ng-click="vector.applyPreset()"
               title="Apply the layer style preset">
